+++ src/components/AuthenticatedRoute/index.tsx
@@ -15,16 +15,11 @@ interface AuthenticatedRouteProps extends OwnProps, StateProps {
 }
 
 const AuthenticatedRoute: React.FC<AuthenticatedRouteProps> = ({ children, tokenData, ...rest }) => {
+
     return (
-        <>
-            {tokenData ? (
-                <Route {...rest} >
-                    {children}
-                </Route>
-            ) : (
-                <Redirect to="/splash" />
-            )}
-        </>
+        <Route {...rest}>
+            {tokenData ? children :  <Redirect to="/splash" />}
+        </Route>
     )
 }
