+++ src/components/Forms/SignInForm/index.test.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import SignInForm from './index';
+import { shallow } from 'enzyme';
+
+test('renders SignInForm without crashing', () => {
+    const baseElement = shallow(
+        <SignInForm/>
+    );
+    expect(baseElement).toBeDefined();
+});
